Use hook for first / top supporter logic
