+++ packages/mobile/src/screens/tip-artist-screen/SendTipScreen.tsx
@@ -1,12 +1,6 @@
 import { useCallback, useEffect, useState } from 'react'
 
-import { ID } from 'audius-client/src/common/models/Identifiers'
-import { Supporter } from 'audius-client/src/common/models/Tipping'
-import {
-  BNWei,
-  StringAudio,
-  StringWei
-} from 'audius-client/src/common/models/Wallet'
+import { BNWei, StringWei } from 'audius-client/src/common/models/Wallet'
 import { getAccountUser } from 'audius-client/src/common/store/account/selectors'
 import {
   getOptimisticSupporters,
@@ -18,12 +12,8 @@ import {
   fetchUserSupporter
 } from 'audius-client/src/common/store/tipping/slice'
 import { getAccountBalance } from 'audius-client/src/common/store/wallet/selectors'
-import { parseWeiNumber } from 'audius-client/src/common/utils/formatUtil'
-import { Nullable } from 'audius-client/src/common/utils/typeUtils'
-import {
-  parseAudioInputToWei,
-  stringWeiToBN
-} from 'audius-client/src/common/utils/wallet'
+import { stringWeiToBN } from 'audius-client/src/common/utils/wallet'
+import { useGetSupport } from 'audius-client/src/hooks/useGetSupport'
 import BN from 'bn.js'
 
 import IconArrow from 'app/assets/images/iconArrow.svg'
@@ -58,11 +48,6 @@ const useStyles = makeStyles(({ spacing }) => ({
 
 const zeroWei = stringWeiToBN('0' as StringWei)
 
-const parseToBNWei = (tipAmount: StringAudio) => {
-  if (!tipAmount) return zeroWei
-  return parseAudioInputToWei(tipAmount) as BNWei
-}
-
 export const SendTipScreen = () => {
   const styles = useStyles()
   const [tipAmount, setTipAmount] = useState('')
@@ -75,34 +60,24 @@ export const SendTipScreen = () => {
   const supportersMap = useSelectorWeb(getOptimisticSupporters)
   const supportingMap = useSelectorWeb(getOptimisticSupporting)
   const receiver = useSelectorWeb(getSendUser)
-  const [
+
+  const {
     amountToTipToBecomeTopSupporter,
-    setAmountToTipToBecomeTopSupporter
-  ] = useState<Nullable<BNWei>>(null)
-  const [supportingAmount, setSupportingAmount] = useState<Nullable<StringWei>>(
-    null
-  )
-  const [topSupporter, setTopSupporter] = useState<Nullable<Supporter>>(null)
-  const [isFirstSupporter, setIsFirstSupporter] = useState(false)
+    shouldFetchUserSupporter,
+    isFirstSupporter,
+    tipAmountWei,
+    hasInsufficientBalance
+  } = useGetSupport({
+    tipAmount,
+    accountBalance,
+    account,
+    receiver,
+    supportingMap,
+    supportersMap
+  })
 
-  const tipAmountWei = parseToBNWei(tipAmount)
-  const hasInsufficientBalance = tipAmountWei.gt(accountBalance)
-
-  /**
-   * Get supporting info if current user is already supporting receiver
-   * so that the already supported amount can be used to determine
-   * how much is left to tip to become top supporter
-   */
   useEffect(() => {
-    if (!account || !receiver) return
-    if (supportingAmount) return
-
-    const supportingForAccount = supportingMap[account.user_id] ?? null
-    const accountSupportingReceiver =
-      supportingForAccount?.[receiver.user_id] ?? null
-    if (accountSupportingReceiver) {
-      setSupportingAmount(accountSupportingReceiver.amount)
-    } else {
+    if (shouldFetchUserSupporter && account && receiver) {
       dispatchWeb(
         fetchUserSupporter({
           currentUserId: account.user_id,
@@ -111,68 +86,7 @@ export const SendTipScreen = () => {
         })
       )
     }
-  }, [account, receiver, supportingAmount, supportingMap, dispatchWeb])
-
-  /**
-   * Get user who is top supporter to later check whether it is
-   * not the same as the current user
-   */
-  useEffect(() => {
-    if (!receiver) return
-
-    const supportersForReceiver = supportersMap[receiver.user_id] ?? {}
-    const rankedSupportersList = ((Object.keys(
-      supportersForReceiver
-    ) as unknown) as ID[])
-      .sort(
-        (k1, k2) =>
-          supportersForReceiver[k1].rank - supportersForReceiver[k2].rank
-      )
-      .map(k => supportersForReceiver[k])
-    const theTopSupporter =
-      rankedSupportersList.length > 0 ? rankedSupportersList[0] : null
-
-    if (theTopSupporter) {
-      setIsFirstSupporter(false)
-      setTopSupporter(theTopSupporter)
-    } else {
-      setIsFirstSupporter(true)
-    }
-  }, [receiver, supportersMap])
-
-  /**
-   * Check whether or not to display prompt to become top or first supporter
-   */
-  useEffect(() => {
-    if (hasInsufficientBalance || !account || !topSupporter) return
-
-    const isAlreadyTopSupporter = account.user_id === topSupporter.sender_id
-    if (isAlreadyTopSupporter) return
-
-    const topSupporterAmountWei = stringWeiToBN(topSupporter.amount)
-    const oneAudioToWeiBN = parseWeiNumber('1') as BNWei
-    let newAmountToTipToBecomeTopSupporter = topSupporterAmountWei.add(
-      oneAudioToWeiBN
-    ) as BNWei
-    if (supportingAmount) {
-      const supportingAmountWei = stringWeiToBN(supportingAmount)
-      newAmountToTipToBecomeTopSupporter = newAmountToTipToBecomeTopSupporter.sub(
-        supportingAmountWei
-      ) as BNWei
-    }
-    if (
-      accountBalance.gte(newAmountToTipToBecomeTopSupporter) &&
-      newAmountToTipToBecomeTopSupporter.gte(oneAudioToWeiBN)
-    ) {
-      setAmountToTipToBecomeTopSupporter(newAmountToTipToBecomeTopSupporter)
-    }
-  }, [
-    hasInsufficientBalance,
-    account,
-    topSupporter,
-    supportingAmount,
-    accountBalance
-  ])
+  }, [shouldFetchUserSupporter, account, receiver, dispatchWeb])
 
   const handleBack = useCallback(() => {
     navigation.goBack()
